backend/requirements.txt
streamlit
psycopg2-binary
streamlit-cookies-manager